Safari-611.1.10/LayoutTests/css3/flexbox/flex-item-child-overflow.html

<!DOCTYPE html>
<html>
<style>
body {
    margin: 0;
}
.flexbox {
    display: flex;
    position: relative;
    flex-flow: column;
}
.flexbox > div {
    height: 0;
}
.flexbox > div > div {
    width: 20px;
    height: 20px;
    line-height: 0px;
    background-color: blue;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

<div data-expected-height=0 data-offset-x=0 data-offset-y=0 class="flexbox">
    <div data-expected-height=0 data-offset-x=0 data-offset-y=0>
        <div data-expected-height=20 data-offset-x=0 data-offset-y=0></div>
    </div>
</div>

</body>
</html>
